+++ CHANGELOG.md
@@ -2,20 +2,25 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-3398] Support classification of data structures
+
 ### Changed
 
-- Timestamps on `DataStructureVersion` are now `utc_datetime_usec`.
+- Timestamps on `DataStructure`, `DataStructureVersion`, `DataStructureRelation`
+  and `StructureMetadata` are now `utc_datetime_usec`.
 
 ## [4.19.0] 2021-05-04
 
 ### Changed
 
 - [TD-3526] Merged `td-dq` with `td-dd`. See `CHANGELOG-dq.md` for changes in
-  `td-dq` previous to this merge.
+  `td-dq` previous to this merge
 - [TD-3621] Increase maximum length for JSON request bodies. The value for JSON
   request bodies can now be configured using the `MAX_PAYLOAD_LENGTH`
   environment variable.
-- [TD-3596] Creates Data Structure Tags
+- [TD-3596] Support tagging of data structures
 
 ### Added